+++ tests/serialization/test_surrogate_serialization.py
@@ -1,6 +1,7 @@
 """Surrogate serialization tests."""
 
 import pytest
+from hypothesis import given
 
 from baybe._optional.info import NGBOOST_INSTALLED, ONNX_INSTALLED
 from baybe.surrogates.base import Surrogate
@@ -10,10 +11,18 @@ from baybe.surrogates.gaussian_process.core import GaussianProcessSurrogate
 from baybe.surrogates.ngboost import NGBoostSurrogate
 from baybe.surrogates.random_forest import RandomForestSurrogate
 from baybe.utils.basic import get_subclasses
+from tests.hypothesis_strategies.surrogates import gaussian_process_surrogates
 from tests.serialization.utils import assert_roundtrip_consistency
 
 
-@pytest.mark.parametrize("surrogate_cls", get_subclasses(Surrogate))
+@pytest.mark.parametrize(
+    "surrogate_cls",
+    [
+        c
+        for c in get_subclasses(Surrogate)
+        if not issubclass(c, GaussianProcessSurrogate)
+    ],
+)
 def test_surrogate_roundtrip(request, surrogate_cls: type[Surrogate]):
     """A serialization roundtrip yields an equivalent object."""
     if issubclass(surrogate_cls, CustomONNXSurrogate):
@@ -25,6 +34,12 @@ def test_surrogate_roundtrip(request, surrogate_cls: type[Surrogate]):
     else:
         surrogate = surrogate_cls()
 
+    assert_roundtrip_consistency(surrogate)
+
+
+@given(gaussian_process_surrogates())
+def test_gaussian_process_surrogate_roundtrip(surrogate: GaussianProcessSurrogate):
+    """A serialization roundtrip yields an equivalent GP surrogate for all configs."""
     assert_roundtrip_consistency(surrogate)